# internal/web/handlers/api/openapi.yaml
openapi: 3.1.0
info:
  title: Opengist API
  version: "1.0.0"
  description: |
    REST API for Opengist.
    
    The base URL is `/api/v1`.

    Authentication uses a Personal Access Token sent in the `Authorization`
    header using the `Bearer` scheme:

        Authorization: Bearer og_xxxxxxxx

    Example usage:
    
        curl -H "Authorization: Bearer og_xxxxxxxx" https://opengist.local/api/v1/gists

    Tokens are created from the web UI under Settings → Access Tokens and carry
    per-resource scopes (gist, user) at read or read/write level.
  license:
    name: AGPL-3.0
    url: https://github.com/thomiceli/opengist/blob/master/LICENSE
servers:
  - url: /api/v1
    description: Current instance
security:
  - bearerAuth: []
...
paths:
...
  /gists/{uuid}:
    parameters:
      - { $ref: "#/components/parameters/GistUuidParam" }
    get:
      operationId: getGist
      summary: Get a gist
      tags: [Gists]
      security:
        - bearerAuth: []
        - {} # Authorization optional (anonymous allowed)
      description: |
        Returns a single gist with its file contents, commits and forks.

        Public and unlisted gists are readable by anyone, including anonymous
        callers. A private gist resolves only for its owner using a token with
        the `gist:read` scope; every other caller receives `404` so the gist's
        existence is not revealed.
      responses:
        "200":
          description: The gist.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Gist" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateGist
      summary: Update a gist
      tags: [Gists]
      security:
        - bearerAuth: ["gist:write"]
      description: |
        **Required scope:** `gist:write`

        Updates a gist owned by the authenticated caller. Only fields present in
        the body are changed; omitted fields are left untouched, and at least
        one of `title`,
        `description`, `visibility` or `files` must be supplied.

        Each entry in `files` is keyed by the current filename and may:

        - set `content` to replace the file body;
        - set `filename` to rename the file;
        - set both to do both at once;
        - be `null` (or carry neither field) to delete the file.

        A key that doesn't match an existing filename adds a new file, in which
        case `content` is required.

        A private gist resolves only for its owner with a `gist:read`-capable
        token; other callers get `404`. A visible gist the caller doesn't own
        returns `403`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GistUpdateRequest" }
      responses:
        "200":
          description: The updated gist.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Gist" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    delete:
      operationId: deleteGist
      summary: Delete a gist
      tags: [Gists]
      security:
        - bearerAuth: ["gist:write"]
      description: |
        **Required scope:** `gist:write`

        Deletes a gist owned by the authenticated caller, removing both its
        repository and database row.

        A private gist resolves only for its owner with a `gist:read`-capable
        token; other callers get `404`. A visible gist the caller doesn't own
        returns `403`.
      responses:
        "204":
          description: The gist was deleted.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /user/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: The user's numeric ID.
        schema: { type: integer }
    get:
      operationId: getUserById
      summary: Get a user by ID
      tags: [Users]
      security:
        - bearerAuth: []
        - {} # Authorization optional (anonymous allowed)
      description: |
        Looks up a user by numeric ID and returns the public user shape (no
        private fields). Readable by anyone, including anonymous callers.
      responses:
        "200":
          description: The user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleUser" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Personal Access Token. Send as `Authorization: Bearer og_<token>`.

        Tokens carry per-resource scopes, granted at read or read/write level
        from the web UI (Settings → Access Tokens). Each operation lists the
        scope it requires in its `security` entry, e.g. `bearerAuth: [gist:write]`.
        An empty list (`bearerAuth: []`) means any valid token is accepted, with
        no specific scope required. The available scopes are:

        - `gist:read` - read gists (including the caller's private/unlisted ones).
        - `gist:write` - create, update, delete and fork gists.
        - `user:read` - read the authenticated user's account.
        - `user:write` - update the authenticated user and toggle likes.
  parameters:
...
    GistUuidParam:
      name: uuid
      in: path
      required: true
      description: The gist's UUID.
      schema: { type: string }
...
  schemas:
    Error:
      type: object
      properties:
        message:
          description: Human-readable error message.
          type: string
        status:
          description: HTTP status code, repeated in the body.
          type: integer
      required: [message, status]
    SimpleUser:
      type: object
      description: |
        The public user shape, embedded as the `owner` of a gist and returned
        by the public user-lookup endpoints. Carries no private fields.
      properties:
        id: { type: integer }
        login: { type: string }
        username: { type: string }
        avatar_url: { type: string, format: uri }
        type: { type: string }
        created_at: { type: string, format: date-time }
...
    GistSimple:
      type: object
      description: |
        The list-shape gist returned by collection endpoints. Does not include
        file contents; fetch the gist detail or raw files for those.
      properties:
        id: { type: string }
        slug_url: { type: string }
        owner: { $ref: "#/components/schemas/SimpleUser" }
        title: { type: string }
        html_url: { type: string, format: uri }
        description: { type: string }
        public:
          type: boolean
          description: True only when `visibility` is `public`.
        visibility:
          type: string
          enum: [public, unlisted, private]
        like_count: { type: integer }
        fork_count: { type: integer }
        clone_url: { type: string, format: uri }
        ssh_url: { type: string }
        topics:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    GistFile:
      type: object
      description: One file of a gist, as returned in gist detail responses.
      properties:
        filename: { type: string }
        type: { type: string }
        language: { type: string }
        size: { type: integer }
        truncated: { type: boolean }
        content: { type: string }
        encoding: { type: string }
    CommitAuthor:
      type: object
      description: Raw git author metadata for a commit.
      properties:
        name: { type: string }
        email: { type: string }
    CommitChangeStatus:
      type: object
      description: Shortstat breakdown for a commit; total equals additions + deletions.
      properties:
        files_changed: { type: integer }
        additions: { type: integer }
        deletions: { type: integer }
        total: { type: integer }
    GistCommit:
      type: object
      properties:
        version:
          type: string
          description: Commit SHA.
        author: { $ref: "#/components/schemas/CommitAuthor" }
        user:
          description: Opengist account matching the commit email, or null.
          oneOf:
            - { $ref: "#/components/schemas/SimpleUser" }
            - { type: "null" }
        change_status: { $ref: "#/components/schemas/CommitChangeStatus" }
        committed_at: { type: string, format: date-time }
    Gist:
      description: |
        Full gist representation including file contents, commits and forks.
        Returned by the create, get and update endpoints.
      allOf:
        - { $ref: "#/components/schemas/GistSimple" }
        - type: object
          properties:
            fork_of:
              description: The parent gist when this is a fork, otherwise null.
              oneOf:
                - { $ref: "#/components/schemas/GistSimple" }
                - { type: "null" }
            forks:
              type: array
              items: { $ref: "#/components/schemas/GistSimple" }
            files:
              type: object
              additionalProperties: { $ref: "#/components/schemas/GistFile" }
            commits:
              type: array
              items: { $ref: "#/components/schemas/GistCommit" }
            truncated: { type: boolean }
...
    GistUpdateFileInput:
      type: object
      description: |
        Per-file entry of an update request, keyed by the current filename.
        A `null` entry (or one with neither field set) deletes the file.
      properties:
        content:
          type: string
          description: New file body. Required when the key names a new file.
        filename:
          type: string
          description: New name for the file (rename).
    GistUpdateRequest:
      type: object
      description: |
        Request body for updating a gist. Every field is optional; supply at
        least one of `title`, `description`, `visibility` or `files`.
      properties:
        title: { type: string }
        description: { type: string }
        visibility:
          type: string
          enum: [public, unlisted, private]
        files:
          type: object
          description: Files keyed by their current filename.
          additionalProperties:
            oneOf:
              - { $ref: "#/components/schemas/GistUpdateFileInput" }
              - { type: "null" }
...
  responses:
    BadRequest:
      description: The request was malformed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: The token lacks the required scope or permission.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: The requested resource does not exist (or is not visible to the caller).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ValidationFailed:
      description: The request body failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }